# codegen-test/swagger.json
{"swagger":"2.0","paths":{"/":{"get":{"operationId":"getAllComments","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/Comment"}},"headers":{}}}},"post":{"operationId":"addMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Comment"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Comment"},"headers":{}}}}},"/application.wadl":{"get":{"operationId":"getWadl","produces":["application/vnd.sun.wadl+xml","application/xml"],"parameters":[],"responses":{"default":{"description":"successful operation"}}}},"/application.wadl/{path}":{"get":{"operationId":"getExternalGrammar","produces":["application/xml"],"parameters":[{"name":"path","in":"path","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/date/{dateString}":{"get":{"operationId":"getRequestedDate","produces":["text/plain"],"parameters":[{"name":"dateString","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"},"headers":{}}}}},"/injectdemo/annotations":{"get":{"operationId":"getParamsUsingAnnotation","consumes":["text/plain"],"produces":["text/plain"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"type":"string"}},{"name":"customHeaderValue","in":"header","required":false,"type":"string"},{"name":"name","in":"cookie","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"},"headers":{}}}}},"/injectdemo/context":{"get":{"operationId":"getParamsUsingContext","consumes":["text/plain"],"produces":["text/plain"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"string"},"headers":{}}}}},"/messages":{"get":{"operationId":"getXmlMessages","consumes":["application/json"],"produces":["text/xml"],"parameters":[{"name":"year","in":"query","required":false,"type":"integer","format":"int32"},{"name":"start","in":"query","required":false,"type":"integer","format":"int32"},{"name":"size","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/Message"}},"headers":{}}}},"post":{"operationId":"addMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Message"}}],"responses":{"default":{"description":"successful operation"}}}},"/messages/{messageId}":{"get":{"operationId":"getMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Message"},"headers":{}}}},"put":{"operationId":"updateMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Message"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Message"},"headers":{}}}},"delete":{"operationId":"deleteMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"default":{"description":"successful operation"}}}},"/messages/{messageId}/comments":{"get":{"operationId":"getAllComments","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/Comment"}},"headers":{}}}},"post":{"operationId":"addMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Comment"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Comment"},"headers":{}}}}},"/messages/{messageId}/comments/{commentId}":{"get":{"operationId":"getMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"name":"commentId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Comment"},"headers":{}}}},"put":{"operationId":"updateMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"name":"commentId","in":"path","required":true,"type":"integer","format":"int64"},{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Comment"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Comment"},"headers":{}}}},"delete":{"operationId":"deleteComment","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"name":"commentId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"default":{"description":"successful operation"}}}},"/myresource":{"get":{"operationId":"getIt","produces":["text/plain"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"string"},"headers":{}}}}},"/profiles":{"get":{"operationId":"getProfiles","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/Profile"}},"headers":{}}}},"post":{"operationId":"addProfile","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Profile"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Profile"},"headers":{}}}}},"/profiles/{profileName}":{"get":{"operationId":"getProfile","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"profileName","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Profile"},"headers":{}}}},"put":{"operationId":"updateProfile","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"profileName","in":"path","required":true,"type":"string"},{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Profile"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Profile"},"headers":{}}}},"delete":{"operationId":"removeProfile","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"profileName","in":"path","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/secured/message":{"get":{"operationId":"securedMethod","produces":["text/plain"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"string"},"headers":{}}}}},"/test":{"get":{"operationId":"testMethod","produces":["text/plain","text/shortdate"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"string","format":"date-time"},"headers":{}}}}},"/{commentId}":{"get":{"operationId":"getMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"name":"commentId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Comment"},"headers":{}}}},"put":{"operationId":"updateMessage","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"name":"commentId","in":"path","required":true,"type":"integer","format":"int64"},{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Comment"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Comment"},"headers":{}}}},"delete":{"operationId":"deleteComment","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"messageId","in":"path","required":true,"type":"integer","format":"int64"},{"name":"commentId","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"default":{"description":"successful operation"}}}}},"definitions":{"Comment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"message":{"type":"string"},"created":{"type":"string","format":"date-time"},"author":{"type":"string"}}},"Message":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"message":{"type":"string"},"created":{"type":"string","format":"date-time"},"author":{"type":"string"},"comments":{"type":"object","additionalProperties":{"$ref":"#/definitions/Comment"}},"links":{"type":"array","items":{"$ref":"#/definitions/Link"}}}},"CommentResource":{"type":"object"},"Profile":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"profileName":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"created":{"type":"string","format":"date-time"}}},"Link":{"type":"object","properties":{"link":{"type":"string"},"rel":{"type":"string"}}}}}