# src/main/java/org/syachiku/madao/messenger/model/Message.java
package org.syachiku.madao.messenger.model;

import java.util.ArrayList;
import java.util.Date;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

import javax.xml.bind.annotation.XmlRootElement;
import javax.xml.bind.annotation.XmlTransient;

/**
 * Message model of the service.
 * 
 * @see Comment
 * @see Profile
 * @author Ellis
 * @since 2016-05-19
 */

@XmlRootElement
public class Message {
	
	private long id;
	private String message;
	private Date created;
	private String author;
	private Map<Long, Comment> comments = new HashMap<>();
	List<Link> links = new ArrayList<>();

	public Message(){
	}
	
	/**
	 * Constructs a new Message.
	 * 
	 * @param id long number to indentify the Message
	 * @param message string content of the Message
	 * @param author string of the author's name
	 */
	public Message(long id, String message, String author) {
		this.id = id;
		this.message = message;
		this.author = author;
		this.created = new Date();
	}
	
	public long getId() {
		return id;
	}
	public void setId(long id) {
		this.id = id;
	}
	public String getMessage() {
		return message;
	}
	public void setMessage(String message) {
		this.message = message;
	}
	public Date getCreated() {
		return created;
	}
	public void setCreated(Date created) {
		this.created = created;
	}
	public String getAuthor() {
		return author;
	}
	public void setAuthor(String author) {
		this.author = author;
	}
	
	@XmlTransient
	public Map<Long, Comment> getComments() {
		return comments;
	}
	public void setComments(Map<Long, Comment> comments) {
		this.comments = comments;
	}

	public List<Link> getLinks() {
		return links;
	}

	public void setLinks(List<Link> links) {
		this.links = links;
	}
	
	/**
	 * Adds link in this model for HATEOAS model propose.
	 * 
	 * @param url url to be provided in the return message instance.
	 * @param rel The relation of the url is refering to between this message.
	 */
	public void addLink(String url, String rel){
		Link link = new Link();
		link.setLink(url);
		link.setRel(rel);
		links.add(link);
	}
}